Completed find & retry single file tests
